requirements.txt
aiohappyeyeballs==2.5.0
aiohttp==3.11.13
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
async-timeout==5.0.1
attrs==25.1.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.2
frozenlist==1.5.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
markdown-to-mrkdwn==0.1.4
multidict==6.1.0
mypy==1.15.0
mypy-extensions==1.0.0
openai==1.65.4
pillow==11.1.0
propcache==0.3.0
pydantic==2.10.6
pydantic_core==2.27.2
requests==2.32.3
slackclient==2.9.4
sniffio==1.3.1
soupsieve==2.6
tqdm==4.67.1
types-requests==2.32.0.20250306
typing_extensions==4.12.2
urllib3==2.3.0
yarl==1.18.3
